feat: refactored benchmarks and added `CliBenchmarks`
