+++ Src/CSharpier.Benchmarks/XmlBenchmarks.cs
@@ -1,0 +1,75 @@
+using System.Xml;
+using System.Xml.Linq;
+using BenchmarkDotNet.Attributes;
+using CSharpier.Core;
+using CSharpier.Core.Xml;
+
+namespace CSharpier.Benchmarks;
+
+[MemoryDiagnoser]
+public class XmlBenchmarks
+{
+    [Benchmark]
+    public void XmlDocument_Parse()
+    {
+        var root = new XmlDocument();
+        root.LoadXml(this.largeXmlCode);
+    }
+
+    [Benchmark]
+    public void XDocument_Parse()
+    {
+        _ = XDocument.Parse(this.largeXmlCode);
+    }
+
+    [Benchmark]
+    public void CustomParser_Parse()
+    {
+        _ = RawNodeReader.ParseXml(
+            this.largeXmlCode,
+            Environment.NewLine,
+            XmlWhitespaceSensitivity.Strict
+        );
+    }
+
+    [Benchmark]
+    public void XmlReader_Parse()
+    {
+        using var xmlReader = XmlReader.Create(
+            new StringReader(this.largeXmlCode),
+            new XmlReaderSettings { IgnoreWhitespace = false }
+        );
+
+        while (xmlReader.Read())
+        {
+            //
+        }
+    }
+
+    private readonly string largeXmlCode = File.ReadAllText(
+        Path.Combine(RepoRoot, "Src/CSharpier.BenchMarks/CodeSamples/Type.xml")
+    );
+
+    // benchmarks creates a new build inside of the bin and doesn't seem to respect CopyToOutputDirectory
+    // finding the .git folder seems like the best way of dealing with the path to the code files
+    private static string RepoRoot { get; } = GetRepoRoot();
+
+    private static string GetRepoRoot()
+    {
+        var currentDirectory = new DirectoryInfo(Directory.GetCurrentDirectory());
+        while (
+            currentDirectory != null
+            && !Directory.Exists(Path.Combine(currentDirectory.FullName, ".git"))
+        )
+        {
+            currentDirectory = currentDirectory.Parent;
+        }
+
+        if (currentDirectory is null)
+        {
+            throw new Exception("Could not find .git directory");
+        }
+
+        return currentDirectory.FullName;
+    }
+}
